+++ src/android/app/src/main/java/org/citra/citra_emu/dialogs/RunningSettingDialog.java
@@ -520,11 +520,10 @@ public class RunningSettingDialog extends DialogFragment {
                     isChanged = true;
                 }
             }
-            // apply settings changes
+            // apply settings changed
             if (isChanged) {
                 NativeLibrary.setRunningSettings(newSettings);
             }
         }
     }
-
 }
