+++ src/android/app/src/main/res/drawable/outline_drive_file_rename_outline_24.xml
@@ -1,0 +1,16 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:tint="?attr/colorControlNormal"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M15,16l-4,4l10,0l0,-4z" />
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M12.06,7.19L3,16.25V20h3.75l9.06,-9.06L12.06,7.19zM5.92,18H5v-0.92l7.06,-7.06l0.92,0.92L5.92,18z" />
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M18.71,8.04c0.39,-0.39 0.39,-1.02 0,-1.41l-2.34,-2.34C16.17,4.09 15.92,4 15.66,4c-0.25,0 -0.51,0.1 -0.7,0.29l-1.83,1.83l3.75,3.75L18.71,8.04z" />
+</vector>
