+++ src/android/app/src/main/res/layout/dialog_editor.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_margin="0dp"
+    android:orientation="vertical"
+    android:padding="0dp">
+
+    <EditText
+        android:id="@+id/setting_editor"
+        android:layout_width="match_parent"
+        android:layout_height="80dp"
+        android:layout_marginStart="6dp"
+        android:layout_marginEnd="6dp"
+        android:autofillHints=""
+        android:gravity="top|start"
+        android:inputType="textMultiLine"
+        android:scrollbars="horizontal|vertical"
+        android:textSize="16sp"
+        android:typeface="monospace"
+        tools:ignore="LabelFor" />
+
+</RelativeLayout>
