+++ src/android/app/src/main/res/menu/menu_file_picker.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/menu_current_directory"
+        android:icon="@drawable/outline_drive_file_rename_outline_24"
+        android:title="@string/current_directory"
+        app:showAsAction="always" />
+
+</menu>
